deadlines op

+++ src/AirShuttle/ServicesPlanner.h
@@ -45,7 +45,7 @@ public:
 	
 	std::vector<Edge> calculatePath(const std::set<Vertex*>& reservations);
 	void planVansFleetMixingPassengers();
-	int objectiveFunction();
+	long int objectiveFunction();
 	void integrateClientWithNoReservation(const Reservation & res, Time waitingTime);
 	void sameDestIntegration(const Reservation & reservation, Service & service);
 	void newDestIntegration(const Reservation & reservation, Service & service);
